requirements.txt
# Web framework
Flask>=2.0

# Image handling
Pillow>=9.0

# Video/Audio handling and metadata
ffmpeg-python>=0.2
mutagen>=1.45

# JWT Authentication
PyJWT>=2.1

# For type hints and ABC
typing-extensions>=4.0

# For data serialization (if needed, e.g., human-friendly YAML config)
PyYAML>=6.0

# For Cloud Storage (optional, uncomment as needed)
# boto3>=1.21      # AWS S3
# google-cloud-storage>=2.0  # Google Cloud Storage
# azure-storage-blob>=12.8   # Azure Blob Storage
